+++ .gitignore
@@ -19,3 +19,4 @@ a.out
 skills-lock.json
 src/markdown.d.ts
 e2e/_test-import*.spec.ts
+e2e/auth-utils-edge.spec.ts
